chore: address quality scorecard findings #533 #534 #535
- #533 (security): deserialize() now requires an explicit trusted=True
  opt-in before unpickling; refuses with CvxError otherwise. Guards the
  pickle.load RCE trust boundary.
- #534 (complexity): extract _Problem container + serialize/deserialize
  round-trip into cvxmarkowitz.problem; builder.py MI 68.26 -> 78.10.
  deserialize/_Problem/CvxError re-exported from builder for compat.
- #535 (test design): add mutation-killing tests — non-square Cholesky
  shape checks in factor/sample risk models (pins shape[0]) and a
  _Problem immutability test. Mutation score 67.8% -> 69.6%.
